Вынес иконки пинов в utils.js
