+++ project/src/components/ui/map/useMap.js
@@ -2,7 +2,6 @@ import {useEffect, useState} from 'react';
 import leaflet from 'leaflet';
 import 'leaflet/dist/leaflet.css';
 import {MapSetting} from './constants.js';
-import {Icon} from './constants';
 
 function useMap(mapRef, city) {
   const [map, setMap] = useState(null);
@@ -29,20 +28,7 @@ function useMap(mapRef, city) {
 
   }, [mapRef, map, city]);
 
-  const icon = leaflet.icon({
-    iconUrl: Icon.REGULAR.URL,
-    iconSize: Icon.REGULAR.SIZE,
-    iconAnchor: Icon.REGULAR.ANCHOR,
-  });
-
-  const iconActive = leaflet.icon({
-    iconUrl: Icon.ACTIVE.URL,
-    iconSize: Icon.ACTIVE.SIZE,
-    iconAnchor: Icon.ACTIVE.ANCHOR,
-  });
-
-
-  return {map, icon, iconActive};
+  return map;
 }
 
 export default useMap;
